// app/Application/Types/PetsData.ts
import React from 'react'

export type UsePetsData = {
    pet_id: number;
    Pet_Name: string;
    Sex: string;
    Age: string;
    Breed: string;
    Color: string;
    image: string;
    Microchip_Number: number;
    Neutered_Spay: string;
    Special_Markings: string;
    Species: string;
    Weight: number;
  };
  
const PetsData = () => {
  return 
}

export default PetsData